pyproject.toml
[build-system]
requires = ["setuptools>=61.2", "setuptools_scm[toml]>=3.4.1"]
build-backend = "setuptools.build_meta"

[project]
name = "zipp"
authors = [
	{ name = "Jason R. Coombs", email = "jaraco@jaraco.com" },
]
description = "Backport of pathlib-compatible object wrapper for zip files"
readme = "README.rst"
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Intended Audience :: Developers",
	"License :: OSI Approved :: MIT License",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3 :: Only",
]
requires-python = ">=3.8"
dependencies = [
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/jaraco/zipp"

[project.optional-dependencies]
test = [
	# upstream
	"pytest >= 6, != 8.1.*",
	"pytest-checkdocs >= 2.4",
	"pytest-cov",
	"pytest-mypy",
	"pytest-enabler >= 2.2",
	"pytest-ruff >= 0.2.1",

	# local
	"jaraco.itertools",
	"jaraco.functools",
	"more_itertools",
	"big-O",
	"pytest-ignore-flaky",
	"jaraco.test",
	"importlib_resources; python_version < '3.9'",
]
doc = [
	# upstream
	"sphinx >= 3.5",
	"jaraco.packaging >= 9.3",
	"rst.linker >= 1.9",
	"furo",
	"sphinx-lint",

	# tidelift
	"jaraco.tidelift >= 1.4",

	# local
]

[tool.setuptools_scm]
